// server/src/api/@generated/user/user-max-order-by-aggregate.input.ts
import { Field } from '@nestjs/graphql';
import { InputType } from '@nestjs/graphql';
import { SortOrder } from '../prisma/sort-order.enum';

@InputType()
export class UserMaxOrderByAggregateInput {
  @Field(() => SortOrder, { nullable: true })
  id?: keyof typeof SortOrder;

  @Field(() => SortOrder, { nullable: true })
  createdAt?: keyof typeof SortOrder;

  @Field(() => SortOrder, { nullable: true })
  updatedAt?: keyof typeof SortOrder;

  @Field(() => SortOrder, { nullable: true })
  deletedAt?: keyof typeof SortOrder;

  @Field(() => SortOrder, { nullable: true })
  lastSeen?: keyof typeof SortOrder;

  @Field(() => SortOrder, { nullable: true })
  disabled?: keyof typeof SortOrder;

  @Field(() => SortOrder, { nullable: true })
  displayName?: keyof typeof SortOrder;

  @Field(() => SortOrder, { nullable: true })
  email?: keyof typeof SortOrder;

  @Field(() => SortOrder, { nullable: true })
  avatarUrl?: keyof typeof SortOrder;

  @Field(() => SortOrder, { nullable: true })
  locale?: keyof typeof SortOrder;

  @Field(() => SortOrder, { nullable: true })
  phoneNumber?: keyof typeof SortOrder;

  @Field(() => SortOrder, { nullable: true })
  passwordHash?: keyof typeof SortOrder;

  @Field(() => SortOrder, { nullable: true })
  emailVerified?: keyof typeof SortOrder;
}